front/requirements.txt
Flask==1.1.2
requests==2.25.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6
WTForms==2.3.3
flask-wtf==0.15.1
python-dateutil
wtforms[email]